Ignore warnings when doing completions.
If a user wants to get the completions, it can be annoying if we fail because of some warning.

Even if the user wants all warnings to be errors when syntax checking, they are unlikely to want
warning to prevent them from writing the code.

So only in the completion execution, add -w to the options to ignore completion.